refactor: get_flash_attn in dedicated file
this is needed to avoid circular dependencies now that we want to get
the flash_attn_version directly in platforms/cuda.py to check if fp8
flash_attn is actually available
